pyproject.toml
[project] # https://packaging.python.org/en/latest/specifications/pyproject-toml/
name = "imperiumengine"
version = "0.0.0"
description = "add"
readme = "README.md"
authors = [{ name = "luiz correa ", email = "zcorrea357@gmail.com" }]
requires-python = ">=3.10,<4.0"
dependencies = [
  "bravado>=11.1.0",
  "cairosvg>=2.7.1",
  "datasets>=3.3.2",
  "griffe-typingdoc>=0.2.8",
  "httpx>=0.28.1",
  "langdetect>=1.0.9",
  "markdown-include-variants>=0.0.4",
  "mdx-include>=1.4.2",
  "mike>=2.1.3",
  "mkdocs-bootswatch>=1.1",
  "mkdocs-glightbox>=0.4.0",
  "mkdocs-macros-plugin>=1.3.7",
  "mkdocs-material>=9.6.5",
  "mkdocs-redirects>=1.2.2",
  "mkdocstrings[python]>=0.28.1",
  "numpy>=2.2.3",
  "pandas>=2.2.3",
  "pillow>=11.1.0",
  "prophet>=1.1.6",
  "python-binance>=1.0.27",
  "pytkdocs>=0.16.2",
  "pyyaml>=6.0.2",
  "requests>=2.32.3",
  "toml>=0.10.2",
  "typer>=0.15.1",
  "types-requests>=2.32.0.20241016",
  "types-toml>=0.10.8.20240310",
]

[project.urls] # https://packaging.python.org/en/latest/specifications/well-known-project-urls/#well-known-labels
homepage = "https://github.com/ImperiumSyndicate/ImperiumEngine"
source = "https://github.com/ImperiumSyndicate/ImperiumEngine"
changelog = "https://github.com/ImperiumSyndicate/ImperiumEngine/blob/main/CHANGELOG.md"
releasenotes = "https://github.com/ImperiumSyndicate/ImperiumEngine/releases"
documentation = "https://github.com/ImperiumSyndicate/ImperiumEngine"
issues = "https://github.com/ImperiumSyndicate/ImperiumEngine/issues"

[dependency-groups] # https://docs.astral.sh/uv/concepts/projects/dependencies/#development-dependencies
dev = [
  "commitizen (>=4.1.0)",
  "coverage[toml] (>=7.6.10)",
  "ipykernel (>=6.29.4)",
  "ipython (>=8.18.0)",
  "ipywidgets (>=8.1.2)",
  "pdoc (>=15.0.1)",
  "poethepoet (>=0.32.1)",
  "pre-commit (>=4.0.1)",
  "pytest (>=8.3.4)",
  "pytest-mock (>=3.14.0)",
  "pytest-xdist (>=3.6.1)",
  "ruff (>=0.9.2)",
  "typeguard (>=4.4.1)",
]

[tool.commitizen] # https://commitizen-tools.github.io/commitizen/config/
bump_message = "bump: v$current_version → v$new_version"
tag_format = "v$version"
update_changelog_on_bump = true
version_provider = "pep621"

[tool.coverage.report] # https://coverage.readthedocs.io/en/latest/config.html#report
fail_under = 50
precision = 1
show_missing = true
skip_covered = true

[tool.coverage.run] # https://coverage.readthedocs.io/en/latest/config.html#run
branch = true
command_line = "--module pytest"
data_file = "reports/.coverage"
source = ["src"]

[tool.coverage.xml] # https://coverage.readthedocs.io/en/latest/config.html#xml
output = "reports/coverage.xml"


[tool.pytest.ini_options] # https://docs.pytest.org/en/latest/reference/reference.html#ini-options-ref
addopts = "--color=yes --doctest-modules --exitfirst --failed-first --strict-config --strict-markers --verbosity=2 --junitxml=reports/pytest.xml"
filterwarnings = ["error", "ignore::DeprecationWarning"]
testpaths = ["src", "tests"]
xfail_strict = true

[tool.ruff] # https://docs.astral.sh/ruff/settings/
fix = true
line-length = 100
src = ["src", "tests"]
target-version = "py310"

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = [
  "A",
  "ASYNC",
  "B",
  "BLE",
  "C4",
  "C90",
  "D",
  "DTZ",
  "E",
  "EM",
  "ERA",
  "F",
  "FBT",
  "FLY",
  "FURB",
  "G",
  "I",
  "ICN",
  "INP",
  "INT",
  "ISC",
  "LOG",
  "N",
  "NPY",
  "PERF",
  "PGH",
  "PIE",
  "PL",
  "PT",
  "PTH",
  "PYI",
  "Q",
  "RET",
  "RSE",
  "RUF",
  "S",
  "SIM",
  "SLF",
  "SLOT",
  "T10",
  "T20",
  "TC",
  "TID",
  "TRY",
  "UP",
  "W",
  "YTT",
]
ignore = [
  "D203",
  "S311",
  "UP038",
  "D200",
  'D205',
  'W505',
  "FBT",
  'D102',
  'B',
  'E',
  'TC001',
  'N',
  'PLR',
  'C901',
  'BLE',
  'S106',
  'D414',
  'RUF',
  'D103',
  'EM',
  'TRY',
  'N806',
  'EM101',
  'PLR2004',
  'D401',
  "D101",
  'D400',
  'ERA001',
  "T201",
  "D104",
  "G004",
  "D100",
  "TRY401",
  "D213",
  "E501",
  "RET504",
  "S101",
  "S307",
]
unfixable = ["ERA001", "F401", "F841", "T201", "T203"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.pycodestyle]
max-doc-length = 100

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.uv] # https://docs.astral.sh/uv/reference/settings/
package = true
[tool.poe.tasks] # https://github.com/nat-n/poethepoet

[tool.poe.tasks.docs]
help = "Generate this package's docs using mkdocs"
cmd = "mkdocs build"

[tool.poe.tasks.docs-serve]
help = "Serve the documentation locally"
cmd = "mkdocs serve"

[tool.poe.tasks.lint]
help = "Lint this package"
cmd = """
    pre-commit run
      --all-files
      --color always
    """

[tool.poe.tasks.test]
help = "Test this package"

[[tool.poe.tasks.test.sequence]]
cmd = "coverage run"

[[tool.poe.tasks.test.sequence]]
cmd = "coverage report"

[[tool.poe.tasks.test.sequence]]
cmd = "coverage xml"
